refactor(core): #HttpServerClient
