--- packages/core/src/http/server/http.server.tokens.ts
@@ -1,4 +1,0 @@
-import { createContextToken } from '../../context/context.token.factory';
-import { HttpServer } from '../http.interface';
-
-export const HttpServerClientToken = createContextToken<HttpServer>('HttpServerClientToken');
